% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromsearch.R
\name{chromsearch}
\alias{chromsearch}
\title{Search one chromsome}
\usage{
chromsearch(choose, n, chrom)
}
\arguments{
\item{choose}{List of .bed tables}

\item{n}{Cut-off distance between colocalized sequences}

\item{chrom}{Chromosome to be searched given as e.g. "chr19"}
}
\value{
A table containing the addresses (as one dimensional intervals) of the members of every cluster, with some annotation data.
}
\description{
Search a single chromosome for clusters of TF binding sequences.
Example produces a null result, test the same complex on "chr9"
for a positive reading.
}
\examples{
complex<-list(nfkb1,nfkb2,relb)
chromsearch(complex,150,"chrY")
}
