% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gensearch.R
\name{colocalized_sequential}
\alias{colocalized_sequential}
\title{Sequential cluster search}
\usage{
colocalized_sequential(choose, chr, n)
}
\arguments{
\item{choose}{List of .bed tables}

\item{chr}{Chromosome to be searched given as e.g. "chr19"}

\item{n}{Cut-off distance between colocalized sequences}
}
\value{
Table containing the addresses (as one dimensional intervals) of the members of every cluster, with some annotation data.
}
\description{
Search one chromosome for clusters using default non-parallel processing.
}
\examples{
complex<-list(nfkb1,nfkb2,relb)
colocalized_sequential(complex,"chrY",150)
}
