% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gensearch.R
\name{ColocalizedFullSearch}
\alias{ColocalizedFullSearch}
\title{Colocalized full search.}
\usage{
ColocalizedFullSearch(choose, n, cores)
}
\arguments{
\item{choose}{List of .bed tables}

\item{n}{The cut-off distance}

\item{cores}{Number of cores for parallel processing. Leaving this blank causes the program to use default (series) processing}
}
\value{
List of lists of each cluster found in each chromosome.
}
\description{
Wrapper for \code{colocalized} that searches every chromosome shared
between the given .bed files.
}
