% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gensearch.R
\name{gensearch}
\alias{gensearch}
\title{Whole genome search.}
\usage{
gensearch(choose, n, cores)
}
\arguments{
\item{choose}{List of .bed tables}

\item{n}{The cut-off distance}

\item{cores}{Number of cores for parallel processing. Leaving this blank causes the program to use default (series) processing}
}
\value{
Table containing the addresses (as one dimensional intervals) of the members of every cluster, with some annotation data.
}
\description{
Search the whole genome for clusters of colocalized TF binding sequences.
}
