% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gensearch.R, R/nfkb1.R
\docType{data}
\name{nfkb1}
\alias{nfkb1}
\title{NFKB1 bed file}
\format{A dataframe with 230505 rows and 8 columns}
\source{
\url{https://ccg.epfl.ch/pwmtools/pwmscan.php}

\url{https://ccg.epfl.ch/pwmtools/pwmscan.php}
}
\usage{
nfkb1

nfkb1
}
\description{
nfkb1
}
\author{
Stefan Golas
}
\keyword{datasets}
