% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gensearch.R, R/nfkb2.R
\docType{data}
\name{nfkb2}
\alias{nfkb2}
\title{NFKB2 bed file}
\format{A dataframe with 1901 rows and 8 columns}
\source{
\url{https://ccg.epfl.ch/pwmtools/pwmscan.php}

\url{https://ccg.epfl.ch/pwmtools/pwmscan.php}
}
\usage{
nfkb2

nfkb2
}
\description{
nfkb2
}
\author{
Stefan Golas
}
\keyword{datasets}
