% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gensearch.R, R/relb.R
\docType{data}
\name{relb}
\alias{relb}
\title{RELB bed file}
\format{A dataframe with 1448 rows and 8 columns}
\source{
\url{https://ccg.epfl.ch/pwmtools/pwmscan.php}

\url{https://ccg.epfl.ch/pwmtools/pwmscan.php}
}
\usage{
relb

relb
}
\description{
relb
}
\author{
Stefan Golas
}
\keyword{datasets}
